+++ src/Fluentify/Semantics/IPropertySymbolExtensions.GetDescriptor.cs
@@ -11,11 +11,11 @@ using static Fluentify.DescriptorAttributeGenerator;
 internal static partial class IPropertySymbolExtensions
 {
     /// <summary>
-    /// Gets the descriptor associated with the <paramref name="property"/> when declared within a record.
+    /// Gets the descriptor associated with the <paramref name="property"/>.
     /// </summary>
     /// <param name="property">The property to be checked for the presence of an attribute.</param>
     /// <returns>
-    /// The descriptor associated with the <paramref name="property"/> when declared within a record, otherwise <see langword="null"/>.
+    /// The descriptor associated with the <paramref name="property"/>, otherwise <see langword="null"/>.
     /// </returns>
     public static string? GetDescriptor(this IPropertySymbol property)
     {